Add files via upload
Added functions to frmTopDownAnalysis.cs:
loadColumnConfiguration(): Loads the column configuration of each DataGridView if it exists.  If not, loads the default configuration.  Returns a List<string> to the calling program

configureDataGridView(): Configures the DataGridViews.  If no column configuration is provided for the DataGridView calls the createColumns(DataGridView dgv, bool isSectors = false) function.  Else calls the createColumns(DataGridView dgv, List<string> cols) function.

Created the Note class to be used with each stock.  Stores the date, time, and information for each Note.  Used to hold additional information does not have a column in the database.