Add files via upload
Added functionality to frmNotes.cs.  Adjusted Panel2 of the SplitContainer to be disabled whenever the index of lstNotes changes.  Added an int and DateTime variable, local to the form, to store the index of the specific note and the time a new note was created.  Changed btnSave to btnOk and added a new btnSave for editing and adding notes.  btnSave_Click determines if the Note is new or not and either adds a new note to the list or updates a previous note.

Adjusted the SelectedIndexChanged event of ListBox lstMyCols on frmColumns.cs.  Disabled btnRemove when SYMBOL is selected since that column contains the key for both dictionaries.

Added note functionality to frmIndividualStock.cs.  Opens a new frmNotes.cs window if btnAdd or btnEdit is clicked.  Deletes the selected note if btnDelete is clicked.  Added a populateListBox function to populate the ListBox if OK is received as the DialogResult after clicking btnAdd or btnEdit, and after clicking btnDelete