Add files via upload
Updated DatabaseAccess.cs to include a function to select all the notes contained in the NOTES table and add them to the appropriate Market or Sector

Added the backupNotesXml function to XmlData.cs.  Used to write the XML for each Note in List<Note>.  Called from createBackupXml when iterating through each Market and Sector.

Corrected a small error in frmColumns.cs.