Add files via upload
Added the Preferences class to store the preference information for the program.  Preference code removed from frmTopDownAnalysis.cs.

Added the Stocks class to replace the Globals, Markets, and Sectors classes.  The only difference between the Markets and Sectors class was the calculation figures which was solved by adding the char type attribute and using two if statements.  Combining all these into one class simplifies the majority of the code for the program.